MAJOR Bug fixes + added speed to computer
Ok, WOW that last version I uploaded was buggy. Lot of stupid mistakes
(this is what happens when you code right when you wake up). Went
through and fixed everything, so please NEVER revert to that version.
Also made the computer a little more efficient. Still has a ways to go
for level 3 and up though